♻️refactor: Update data types to use DOUBLE for bill and bill detail attributes

+++ models/Bill.js
@@ -16,23 +16,23 @@ const Bill = connection.define('bills', {
         allowNull: false
     },
     bill_all_amount: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false
     },
     bill_receive: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false
     },
     bill_change: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false
     },
     bill_all_discount: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false
     },
     bill_all_profit: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DOUBLE,
         allowNull: false
     },
     bill_payment_method: {
